Fix compile issues and version number (which has to be a number).

+++ libtest/cpu.hpp
@@ -36,8 +36,10 @@
 
 #pragma once
 
+#include <cstdlib>
+
 namespace libtest {
 
-size_t number_of_cpus();
+uint32_t number_of_cpus();
 
 } // namespace libtest
